commit
commit

+++ src/main/java/com/kakaopay/membership/barcode/controller/dto/BarcodeIssueReqDto.java
@@ -7,5 +7,5 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PUBLIC)
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 public class BarcodeIssueReqDto {
-    private long userId;
+    private Integer userId;
 }
