+++ src/main/java/com/kakaopay/membership/barcode/controller/dto/BarcodeIssueResDto.java
@@ -3,6 +3,8 @@ package com.kakaopay.membership.barcode.controller.dto;
 import com.kakaopay.membership.barcode.service.dto.BarcodeIssueOutDto;
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
@@ -10,7 +12,7 @@ import lombok.*;
 @Builder
 public class BarcodeIssueResDto {
     private String barcode;
-    private long ownderId;
+    private Integer ownderId;
 
     public static BarcodeIssueResDto fromOutDto(BarcodeIssueOutDto inDto) {
         return BarcodeIssueResDto
