+++ src/main/java/com/kakaopay/membership/barcode/service/dto/BarcodeIssueOutDto.java
@@ -3,6 +3,8 @@ package com.kakaopay.membership.barcode.service.dto;
 import com.kakaopay.membership.barcode.entity.Barcode;
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @ToString
@@ -11,7 +13,7 @@ import lombok.*;
 @Builder
 public class BarcodeIssueOutDto {
     private String barcode;
-    private long ownerId;
+    private Integer ownerId;
     public static BarcodeIssueOutDto fromEntity(Barcode entity) {
         return new BarcodeIssueOutDto(entity.getBarcode(), entity.getOwnerId());
     }
